style(C library): Rename C library module
Rename the C library from QMigratelib.py to lib.py -
QMigrate.core.QMigratelib seemed tautological.
